requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
einops==0.8.0
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
huggingface-hub==0.24.1
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.1
mpmath==1.3.0
networkx==3.3
numpy==2.0.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
packaging==24.1
pika==1.3.2
pillow==10.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
setuptools==72.1.0
six==1.16.0
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.4
transformers==4.43.1
triton==3.0.0
typing_extensions==4.12.2
urllib3==2.2.2
autodistill
autodistill_clip
autodistill_grounded_sam
numpy
opencv-python
pillow
roboflow
supervision
tqdm
pycocotools
scikit-image
